Add BookContainer and some style
