+++ src/App.js
@@ -1,39 +1,19 @@
 import React from 'react';
-import { useSelector, useDispatch } from 'react-redux';
-import { addBook, removeBook } from './redux/books/booksSlice';
-import Button from './components/Button';
-import Book from './components/Book';
+import { Routes, Route } from 'react-router-dom';
+import Navigation from './components/Navigation';
+import './components/assets/style.css';
+import Category from './components/Category';
+import BookContainer from './components/BookContainer';
 
 function App() {
-  const books = useSelector((state) => state.books.books);
-  const dispatch = useDispatch();
-
-  const handleAddBook = () => {
-    const newBook = {
-      item_id: `item${books.length + 1}`,
-      title: 'New Book Title',
-      author: 'New Author Name',
-      category: 'Fiction',
-    };
-    dispatch(addBook(newBook));
-  };
-
-  const handleRemoveBook = (itemId) => {
-    dispatch(removeBook(itemId));
-  };
-
   return (
-    <div className="App">
-      <h1>Bookstore</h1>
-      <div>
-        {books.map((book) => (
-          <div key={book.item_id}>
-            <Book book={book} />
-            <Button type="button" onClick={() => handleRemoveBook(book.item_id)} text="Remove Book" />
-          </div>
-        ))}
-      </div>
-      <Button type="button" onClick={handleAddBook} text="Add Book" />
+    <div className="container">
+      <Navigation />
+      <Routes>
+        <Route path="/" element={<BookContainer />} />
+        <Route path="/categories" element={<Category />} />
+      </Routes>
+
     </div>
   );
 }
